refactor: 예약페이지 컴포넌트 분리 및 코드 개선 (#95)
